unified predictor + catcher fixes
