+++ HGR_CNN/HGR_CNN.pyproj
@@ -22,7 +22,9 @@
     <EnableUnmanagedDebugging>false</EnableUnmanagedDebugging>
   </PropertyGroup>
   <ItemGroup>
+    <Compile Include="config.py" />
     <Compile Include="sim.py" />
+    <Compile Include="simulation_catcher.py" />
     <Compile Include="sim_const.py" />
     <Compile Include="simulation_predictor.py">
       <SubType>Code</SubType>
@@ -37,6 +39,7 @@
     <Compile Include="image_data_manager.py" />
     <Compile Include="predictor_facade.py" />
     <Compile Include="simple_recognizer.py" />
+    <Compile Include="video_catcher.py" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="remoteApi.dll" />
