+++ HGR_CNN/blob_recognizer.py
@@ -1,7 +1,10 @@
 import cv2
 
 def find_blob(image):
-    _, contours, _ = cv2.findContours(image, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_NONE)
+    try:
+        _, contours, _ = cv2.findContours(image, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_NONE)
+    except:
+        contours,_ = cv2.findContours(image.astype("uint8"), cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_NONE)
     if len(contours) == 0:
         return None
     largest_blob = max(contours, key=lambda element: cv2.contourArea(element))
