+++ HGR_CNN/simple_recognizer.py
@@ -22,7 +22,7 @@ class BlobRecognizer():
 
     def __find_blob_binary(self,image):
         try:
-            _, contours, hierarchy_ = cv2.findContours(image.astype("uint8"), cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_NONE)
+            _, contours,_ = cv2.findContours(image.astype("uint8"), cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_NONE)
         except:
             contours,_ = cv2.findContours(image.astype("uint8"), cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_NONE)
         try:
